requirements.txt
paho-mqtt==1.3.1
PyYAML==3.13
